+++ packages/sales-api-service/src/server/routes/recurring-payments.js
@@ -1,9 +1,10 @@
 import {
   dueRecurringPaymentsRequestParamsSchema,
   dueRecurringPaymentsResponseSchema,
-  processRPResultRequestParamsSchema
+  processRPResultRequestParamsSchema,
+  cancelRecurringPaymentRequestParamsSchema
 } from '../../schema/recurring-payments.schema.js'
-import { getRecurringPayments, processRPResult } from '../../services/recurring-payments.service.js'
+import { getRecurringPayments, processRPResult, cancelRecurringPayment } from '../../services/recurring-payments.service.js'
 
 const SWAGGER_TAGS = ['api', 'recurring-payments']
 
@@ -52,6 +53,30 @@ export default [
             200: { description: 'New permission from recurring payment record generated successfully' }
           },
           order: 2
+        }
+      }
+    }
+  },
+  {
+    method: 'GET',
+    path: '/cancelRecurringPayment/{id}',
+    options: {
+      handler: async (request, h) => {
+        const { id } = request.params
+        const result = await cancelRecurringPayment(id)
+        return h.response(result)
+      },
+      description: 'Cancel a recurring payment',
+      tags: SWAGGER_TAGS,
+      validate: {
+        params: cancelRecurringPaymentRequestParamsSchema
+      },
+      plugins: {
+        'hapi-swagger': {
+          responses: {
+            200: { description: 'Recurring payment cancelled' }
+          },
+          order: 3
         }
       }
     }